requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==4.2.13
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.11.0
Pillow
pandas
openpyxl
djangorestframework>=3.15.0
openpyxl
WeasyPrint
